requirements.txt
beanstalkc
django
django-cors-headers
django-ldapdb
djangorestframework
pycrypto
PyJWT
python-ldap
requests

